+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## Changed
-- Removed `injectIntl` from example
+## Added
+- React component for summary
+- Extension point in order to render Coupon app
+- Go to checkout button
+- Interface to be used by Cart app
 
 - **Component** Create the VTEX Store Component _IO Base App_